+++ src/components/checkbox/Label.tsx
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export default styled.label`
+  cursor: pointer;
+  margin-left: 4px;
+`;