+++ src/icons/index.tsx
@@ -1,0 +1,7 @@
+import Check from './Check';
+import Dot from './Dot';
+import DownArrow from './DownArrow';
+import Heart from './Heart';
+import UpArrow from './UpArrow';
+
+export { DownArrow, UpArrow, Heart, Check, Dot };